Add a customer menu for wines
The Customer component will take an array of drinks along with the order
function to generate a menu for the user to pick from. The Customer
component will sort the drink array it receives into smaller drink
arrays based on the color attribute of the drink and store it in state.
Customer will then call the Category component for each color array
resulting in a categorized menu of drinks.

 - Update state to include masterDrinkList

 - Add logic to allow updating of masterDrinkList to reduce stock by
one when a drink is ordered

 - Replace buttons within code with a new CustomButton component