Add an employee page
The Employee component will take an array of drinks along with the
restock and delete functions to display current stock for the user to
interact with. the Employee component will sort the drink array it
receives first by the stock attribute then by the color attribute of the
drink and display it. Employee will call the new Form component to
allow the addition of new drinks, along with having buttons to add stock
to or delete existing drinks from the masterDrinkList.

- Add logic to allow addition of new entries into masterDrinkList

- Add logic to allow updating of masterDrinkList to increase stock by
124 when a keg is added to stock

- Add logic to allow deleting of an entry from masterDrinkList